Added vendors method in AdminEstimateRequestController, added vendors\_save and bid methods in the same controller, and added related routes in web.php; updated EstimateRequest model to include bids relationship; updated User model to include assignRole method; updated custom.css; added tooltip functionality in estimate\_req/index.blade.php; added links section in app.blade.php for admin, manager, and user layouts.

+++ database/migrations/2023_09_06_142004_create_stored_services_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
     {
         Schema::create('stored_services', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('customer_id');
+            $table->unsignedBigInteger('customer_id');
             $table->foreign('customer_id')->references('id')->on('customer_details')->onDelete('cascade');
             $table->string('nick_name')->nullable();
             $table->string('primary')->nullable();
